set up some sorry excuses for tests, but got em started
