+++ swaap/swaapTests/swaapTests.swift
@@ -19,25 +19,8 @@ class SwaapTests: XCTestCase {
 //		// Put teardown code here. This method is called after the invocation of each test method in the class.
 //	}
 
-	func testRangeNormalizing() {
-		let rangeA = 0...10
-		let rangeB = -10...10
-		let rangeC = -5.0...11.5
-
-		XCTAssertEqual(0, rangeA.normalizedIndex(0))
-		XCTAssertEqual(1, rangeA.normalizedIndex(10))
-		XCTAssertEqual(0.2, rangeA.normalizedIndex(2))
-		XCTAssertEqual(-1, rangeA.normalizedIndex(-10))
-
-		XCTAssertEqual(0, rangeB.normalizedIndex(-10))
-		XCTAssertEqual(1, rangeB.normalizedIndex(10))
-		XCTAssertEqual(0.5, rangeB.normalizedIndex(0))
-		XCTAssertEqual(1.5, rangeB.normalizedIndex(20))
-
-		XCTAssertEqual(0, rangeC.normalizedIndex(-5))
-		XCTAssertEqual(1, rangeC.normalizedIndex(11.5))
-		XCTAssertEqual(0.5, rangeC.normalizedIndex(3.25))
-		XCTAssertEqual(1.5, rangeC.normalizedIndex(19.75))
-
+	func testCreateUser() {
+		let authManager = AuthManager()
+		let profileController = ProfileController(authManager: authManager)
 	}
 }
